Analyzing Students' Mental Health
Does going to university in a different country affect your mental health? A Japanese international university surveyed its students in 2018 and published a study the following year that was approved by several ethical and regulatory boards.

The study found that international students have a higher risk of mental health difficulties than the general population, and that social connectedness (belonging to a social group) and acculturative stress (stress associated with joining a new culture) are predictive of depression.

For this project, I will explore the `students` data using SQL and Python to find out if we can come to a similar conclusion for international students and see if the length of stay is a contributing factor.